changed the icons to bootstrap's icons
i made a dictionary of jsx in a new file SocialMediaIcons.js so i could call different social media icons imported from bootstrap since the original icons didn't load the images on my end. In that file, I also made a stateless React component that I would put in place of the image tags for the social medias. For this component, I just returned the icon based on the attribute that was entered into the <Icon/> tag. To get it to enter the right icon, i changed the iconSrc variables in linksData.tsx so they would work.

I also changed the css so that the elements in the links were spaced evenly. I also changed the width of the boxes and the border radius.